test: 完成驱动适配器单元测试收口工作
1.  重构测试代码，移除重复的包级测试辅助函数，将其移入对应业务测试文件
2.  简化gomonkey方法补丁调用，移除不必要的reflect.TypeOf调用
3.  迁移并补全build_task、resource相关的缺失单元测试用例
4.  删除冗余的独立测试文件，将测试逻辑合并到对应业务测试文件中
5.  更新UT覆盖率文档，更新覆盖率统计数据并完成收口检查
