+++ adp/vega/vega-backend/server/driveradapters/router_test.go
@@ -14,7 +14,47 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"vega-backend/common"
+	"vega-backend/interfaces"
+	"vega-backend/worker"
 )
+
+func setGinMode() func() {
+	oldMode := gin.Mode()
+	gin.SetMode(gin.TestMode)
+	return func() {
+		gin.SetMode(oldMode)
+	}
+}
+
+func MockNewRestHandler(
+	appSetting *common.AppSetting,
+	as interfaces.AuthService,
+	cs interfaces.CatalogService,
+	rs interfaces.ResourceService,
+	bts interfaces.BuildTaskService,
+	ds interfaces.DatasetService,
+	cts interfaces.ConnectorTypeService,
+	dts interfaces.DiscoverTaskService,
+	dss interfaces.DiscoverScheduleService,
+	rds interfaces.ResourceDataService,
+	sw *worker.ScheduleWorker,
+) *restHandler {
+	return &restHandler{
+		appSetting: appSetting,
+		as:         as,
+		cs:         cs,
+		rs:         rs,
+		bts:        bts,
+		ds:         ds,
+		cts:        cts,
+		dts:        dts,
+		dss:        dss,
+		rds:        rds,
+		sw:         sw,
+	}
+}
 
 func Test_RestHandler_HealthCheck(t *testing.T) {
 	restoreGinMode := setGinMode()
